Add vote column to comments table.
The script update_votes.sql calculates the values for that column based
on the entries in votes table.

This is an optimization to avoid aggregating the votes in a separate
table every time we retrieve comments.

Also updated examples.sql to use the new full_visible_comments view -
this greatly simplified all queries.