Add ability to use local database for authentication.
Add user_auths table that is used for storing passwords etc.
For simplicity the user_auths table is kept in database even when
it's not used.

Add LocalAuth class for performing the authentication using this
local table.

Provide a way in config file to choose between the two authentication
methods.