New /replies request and saving of replies.
The normal queries for comments now exclude the replies.  These
need to be asked for separately using the /replies request.

Adding of replies is done exactly the same as with normal comments -
through the POST /comments request.

For each comment the GET /comments now also returns parentId and
replyCount fields.