Use latest connect-mysql-session directly from git repo.
Although connect-mysql-session 0.1.3 has been out for over a year,
the latest versions have been marked with node ~ 0.4, but we're
already running on node 0.6 for quite some time now.

Jimmy Bergman has done a fix for it, but it hasn't been merged in:
https://github.com/CarnegieLearning/connect-mysql-session/pull/1

Apparently this whole repo has been abandoned by the author.

So for now I'm pulling in the fix from Jimmy's repo (which too is
over a year old).  Upgrading from 0.1.1 to this version got rid
of a mysterious crash in Sencha projects server.  My hypothesis
is that some dependencies of connect-mysql-session got upgraded
and that lead to problems in 0.1.1, but something has been fixed
in this regard in 0.1.3.

Fuzzy stuff...